Added validation

+++ app/Http/Controllers/ProjectController.php
@@ -24,10 +24,18 @@ class ProjectController extends Controller
     {
         $project = new Project;
 
-        $project->title = $request->title;
-        $project->slug = strtolower(str_replace(' ', '-', $request->title));
-        $project->description = $request->description;
+        // Validate the request
+        $request->validate([
+            'title' => 'required',
+            'description' => 'required',
+            'youtube' => 'nullable|regex:/^[a-zA-Z0-9_-]{11}$/',
+            'git' => 'nullable|url',
+            'live' => 'nullable|url'
+        ]);
 
+        $project->slug = strtolower(str_replace(' ', '-', $request->title));
+
+        $project->fill($request->all());
         $project->save();
 
         return $project;
@@ -35,11 +43,24 @@ class ProjectController extends Controller
 
     public function update(Request $request, Project $project)
     {
+        // Validate the request
+        $request->validate([
+            'title' => 'required',
+            'description' => 'required',
+            'youtube' => 'nullable|regex:/^[a-zA-Z0-9_-]{11}$/',
+            'git' => 'nullable|url',
+            'live' => 'nullable|url'
+        ]);
+
         return $project->update($request->all());
     }
 
     public function addMember(Project $project, Member $member)
     {
+        if ($project->members->contains($member->id)) {
+            return response()->json(['error' => 'Member already in project'], 400);
+        }
+
         $project->members()->attach($member->id);
 
         return $this->show($project);
@@ -47,6 +68,10 @@ class ProjectController extends Controller
 
     public function removeMember(Project $project, Member $member)
     {
+        if (!$project->members->contains($member->id)) {
+            return response()->json(['error' => 'Member not in project'], 400);
+        }
+
         $project->members()->detach($member->id);
 
         return $this->show($project);
@@ -54,6 +79,10 @@ class ProjectController extends Controller
 
     public function addTag(Project $project, Tag $tag)
     {
+        if ($project->tags->contains($tag->id)) {
+            return response()->json(['error' => 'Tag already attached to project'], 400);
+        }
+
         $project->tags()->attach($tag->id);
 
         return $this->show($project);
@@ -61,6 +90,9 @@ class ProjectController extends Controller
 
     public function removeTag(Project $project, Tag $tag)
     {
+        if (!$project->tags->contains($tag->id)) {
+            return response()->json(['error' => 'Tag not attached to project'], 400);
+        }
         $project->tags()->detach($tag->id);
 
         return $this->show($project);
